Move `<MetionsSection />` to own file
