--- docs/src/components/Mention/styles.module.css
@@ -1,11 +1,0 @@
-.mention {
-  font-size: 15px;
-}
-
-.mentionMeta {
-  color: var(--ifm-color-emphasis-700);
-}
-
-.mentionMeta strong {
-  color: var(--ifm-font-color-base);
-}
