+++ docs/src/components/MentionsSection/index.js
@@ -5,19 +5,15 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import clsx from 'clsx';
+// @ts-check
 
 import Link from '@docusaurus/Link';
+import MENTIONS from '@site/src/data/mentions';
+import Heading from '@theme/Heading';
+import clsx from 'clsx';
 import styles from './styles.module.css';
 
-export default function Tweet({
-  url,
-  handle,
-  name,
-  content,
-  date,
-  githubUsername,
-}) {
+function Mention({ url, handle, name, content, date, githubUsername }) {
   return (
     <div className={clsx('card', styles.mention)}>
       <div className="card__header">
@@ -43,6 +39,32 @@ export default function Tweet({
         <Link className={clsx(styles.mentionMeta, styles.mentionDate)} to={url}>
           {date}
         </Link>
+      </div>
+    </div>
+  );
+}
+
+export default function MentionsSection() {
+  let columns = [[], [], []];
+  MENTIONS.filter((mention) => mention.showOnHomepage).forEach((mention, i) =>
+    columns[i % 3].push(mention),
+  );
+
+  return (
+    <div className={clsx(styles.section, styles.sectionAlt)}>
+      <div className="container">
+        <Heading as="h2" className={clsx('margin-bottom--lg', 'text--center')}>
+          Loved by many engineers
+        </Heading>
+        <div className={clsx('row', styles.mentionsSection)}>
+          {columns.map((items, i) => (
+            <div className="col col--4" key={i}>
+              {items.map((tweet) => (
+                <Mention {...tweet} key={tweet.url} />
+              ))}
+            </div>
+          ))}
+        </div>
       </div>
     </div>
   );
