+++ docs/src/components/MentionsSection/styles.module.css
@@ -1,0 +1,35 @@
+.mentionsSection > :global(.col) > * {
+  margin-bottom: 2rem;
+}
+
+@media (max-width: 996px) {
+  .mentionsSection > :global(.col:last-child) > *:last-child {
+    margin-bottom: 0;
+  }
+}
+
+@media (min-width: 997px) {
+  .mentionsSection > :global(.col) > *:last-child {
+    margin-bottom: 0;
+  }
+}
+
+.section {
+  padding: 72px 0;
+}
+
+.sectionAlt {
+  background-color: var(--ifm-color-emphasis-100);
+}
+
+.mention {
+  font-size: 15px;
+}
+
+.mentionMeta {
+  color: var(--ifm-color-emphasis-700);
+}
+
+.mentionMeta strong {
+  color: var(--ifm-font-color-base);
+}
