+++ docs/src/pages/index.js
@@ -3,8 +3,7 @@
 import Link from '@docusaurus/Link';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import FeaturesSection from '@site/src/components/FeaturesSection/index';
-import Mention from '@site/src/components/Mention';
-import Mentions from '@site/src/data/mentions';
+import MentionsSection from '@site/src/components/MentionsSection/index';
 import RioLogo from '@site/static/assets/rio-logo.svg';
 import Heading from '@theme/Heading';
 import Layout from '@theme/Layout';
@@ -35,32 +34,6 @@ function HomepageHeader() {
         <RioLogo className={styles.logo} />
       </div>
     </header>
-  );
-}
-
-function MentionsSection() {
-  let columns = [[], [], []];
-  Mentions.filter((mention) => mention.showOnHomepage).forEach((mention, i) =>
-    columns[i % 3].push(mention),
-  );
-
-  return (
-    <div className={clsx(styles.section, styles.sectionAlt)}>
-      <div className="container">
-        <Heading as="h2" className={clsx('margin-bottom--lg', 'text--center')}>
-          Loved by many engineers
-        </Heading>
-        <div className={clsx('row', styles.mentionsSection)}>
-          {columns.map((items, i) => (
-            <div className="col col--4" key={i}>
-              {items.map((tweet) => (
-                <Mention {...tweet} key={tweet.url} />
-              ))}
-            </div>
-          ))}
-        </div>
-      </div>
-    </div>
   );
 }
